requirements.txt
django>=1.4.5
django-tastypie==0.9.14
requests==2.0.1
